% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{adjustNamesEvaluateModel}
\alias{adjustNamesEvaluateModel}
\title{adjustNamesEvaluateModel}
\usage{
adjustNamesEvaluateModel(modelData, modelNames)
}
\arguments{
\item{modelData}{data.frame Evaluate data.}

\item{modelNames}{list Row names to format data.}
}
\value{
data.frame with formated data and respective names.
}
\description{
adjustNamesEvaluateModel
}
