% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-calls.R
\name{apiEvaluateModel}
\alias{apiEvaluateModel}
\title{apiEvaluateModel}
\usage{
apiEvaluateModel(req, filename)
}
\arguments{
\item{req}{list with information about request from plumber API.}

\item{filename}{path to file (Evaluate.OUT) to be parsed.}
}
\value{
data.frame in JSON format with parser result.
}
\description{
apiEvaluateModel
}
