% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-calls.R
\name{apiEvaluateStatistics}
\alias{apiEvaluateStatistics}
\title{apiEvaluateStatistics}
\usage{
apiEvaluateStatistics(req, fileA, fileB)
}
\arguments{
\item{req}{list with information about request from plumber API.}

\item{fileA}{path to file Summary.OUT from DSSAT Crop modelA}

\item{fileB}{path to file Summary.OUT from DSSAT Crop modelA}
}
\value{
data.frame in JSON format with statistics comparison between models
}
\description{
Calculate statistics about compare summary files from two DSSAT crop models
}
