% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-calls.R
\name{apiSummaryDifferences}
\alias{apiSummaryDifferences}
\title{apiSummaryDifferences}
\usage{
apiSummaryDifferences(req, fileA, fileB, tNUM = NULL, tDTA = NULL)
}
\arguments{
\item{req}{list with information about request from plumber API.}

\item{fileA}{path to file Summary.OUT from DSSAT Crop modelA}

\item{fileB}{path to file Summary.OUT from DSSAT Crop modelA}

\item{tNUM}{Threshold for numeric comparison.}

\item{tDTA}{Threshold for date comparison.}
}
\value{
data.frame in JSON format with statistics comparison between models
}
\description{
Calculate statistics about compare summary files from two DSSAT crop models
}
