% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-calls.R
\name{apiSummaryModel}
\alias{apiSummaryModel}
\title{apiSummaryModel}
\usage{
apiSummaryModel(req, filename)
}
\arguments{
\item{req}{list with information about request from plumber API.}

\item{filename}{path to file (Summary.OUT) to be parsed.}
}
\value{
data.frame in JSON format with parser result.
}
\description{
apiSummaryModel
}
