% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{checkColumns}
\alias{checkColumns}
\title{checkColumns}
\usage{
checkColumns(dA, dB)
}
\arguments{
\item{dA}{data.frame with data from modelA}

\item{dB}{data.frame with data from modelB}
}
\value{
list with common variables, extra variables and reference model
}
\description{
Verify two models and check the columns before comparison procedure.
}
