% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{checkFilesSummary}
\alias{checkFilesSummary}
\title{checkFilesSummary}
\usage{
checkFilesSummary(fileA, fileB)
}
\arguments{
\item{fileA}{path to file (Summary.OUT) to be parsed from modelA.}

\item{fileB}{path to file (Summary.OUT) to be parsed from modelB.}
}
\value{
a list with dataset from model A, modelB and warnings
}
\description{
Check if input files (Summary.OUT) from both models could be compared.
}
