% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{checkHeadersSummary}
\alias{checkHeadersSummary}
\title{checkHeadersSummary}
\usage{
checkHeadersSummary(models, tNUM, tDTA)
}
\arguments{
\item{models}{list with models data.frames}

\item{tNUM}{Threshold for numeric comparison.}

\item{tDTA}{Threshold for date comparison.}
}
\value{
a list with dataset from model A, modelB and warnings
}
\description{
Verify data from Summary files.
}
