% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-local-cran.R
\name{createLocalCRAN}
\alias{createLocalCRAN}
\title{createLocalCRAN}
\usage{
createLocalCRAN(base_dir = ".", type = c("both", "win", "posix"))
}
\arguments{
\item{base_dir}{package path directory}

\item{type}{repository type: posix, win or both}
}
\description{
createLocalCRAN
}
