% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{decimalLen}
\alias{decimalLen}
\title{decimalLen}
\usage{
decimalLen(col)
}
\arguments{
\item{col}{numeric vector with column content}
}
\value{
numeric Calculated decimal lenght or NA
}
\description{
Verify if the column type is decimal and calculate it's length
}
