% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-difference-statistics.R
\name{differencesStatisticsSummary}
\alias{differencesStatisticsSummary}
\title{differenceStatisticsSummary}
\usage{
differencesStatisticsSummary(
  fileA = NULL,
  fileB = NULL,
  tNUM = getOption("RCroptest.numericThreshold", default = 2),
  tDTA = getOption("RCroptest.dateThreshold", default = 0)
)
}
\arguments{
\item{fileA}{path to file (Summary.OUT) to be parsed from modelA.}

\item{fileB}{path to file (Summary.OUT) to be parsed from modelB.}

\item{tNUM}{Threshold value for numeric comparison.}

\item{tDTA}{Threshold value for date comparison.}
}
\value{
a list with dataset from model A, modelB and warnings
}
\description{
Check if input files (Summary.OUT) from both models could be compared.
}
