% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluateModel}
\alias{evaluateModel}
\title{Show Evaluate model}
\usage{
evaluateModel(filename)
}
\arguments{
\item{filename}{path to the Evaluate output file from DSSAT-CSM to be parsed.}
}
\value{
A data.frame with the parsed data from evaluate output file.
}
\description{
Show Evaluate model
}
