% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluateStatistics}
\alias{evaluateStatistics}
\title{Show Statistics from comparison between Evaluate outputs.}
\usage{
evaluateStatistics(fileA, fileB)
}
\arguments{
\item{fileA}{path to the summary file from Model A.}

\item{fileB}{path to the summary file from Model B.}
}
\value{
a list  with Crop code and data.frame with results.
}
\description{
Show Statistics from comparison between Evaluate outputs.
}
\examples{

#'  evaluateStatistics('C:/DSSAT47/CTWork/workA/Evaluate.OUT',
#'  'C:/DSSAT47/CTWork/workB/Evaluate.OUT')

}
