% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{makeDiffAt}
\alias{makeDiffAt}
\title{makeDiffAt}
\usage{
makeDiffAt(val1, val2, idx, iDTA)
}
\arguments{
\item{val1}{numeric Value from model A}

\item{val2}{numeric Value from model B}

\item{idx}{numeric Index of the actual column}

\item{iDTA}{numeric Index of Date columns}
}
\value{
a numeric value with the differece between model
}
\description{
Calculate the absolute or relative difference between values
based on the type of column (DATE ou NUMERIC).
}
