% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{modelStatisticsEvaluate}
\alias{modelStatisticsEvaluate}
\title{modelStatisticsEvaluate}
\usage{
modelStatisticsEvaluate(listData, rowNames)
}
\arguments{
\item{listData}{data.frame Data from raw evaluate model.}

\item{rowNames}{list Row names for Statistics data.frame.}
}
\value{
list Processed data.frame info with statistics for use.
}
\description{
Calculate statistics from raw output from Evaluate file.
}
