% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{mRsq}
\alias{mRsq}
\title{mRsq - Calculate R Squared (R2)}
\usage{
mRsq(obs, sim)
}
\arguments{
\item{obs}{observed values vector}

\item{sim}{simulated values vector}
}
\value{
R2 value
}
\description{
mRsq - Calculate R Squared (R2)
}
