% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-output-model.R
\name{outputModelEvaluate}
\alias{outputModelEvaluate}
\title{outputModelEvaluate}
\usage{
outputModelEvaluate(filename)
}
\arguments{
\item{filename}{path to file (Evaluate.OUT) to be parsed.}
}
\value{
a list with model information ID and a data.fraem in JSON format.
}
\description{
outputModelEvaluate
}
