% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-output-model.R
\name{outputModelSummary}
\alias{outputModelSummary}
\title{outputModelSummary}
\usage{
outputModelSummary(filename)
}
\arguments{
\item{filename}{path to file (Summary.OUT) to be parsed.}
}
\value{
a list with model information ID and a data.fraem in JSON format.
}
\description{
outputModelSummary
}
