% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-overall-statistics.R
\name{overallEvaluateSummary}
\alias{overallEvaluateSummary}
\title{overallEvaluateSummary}
\usage{
overallEvaluateSummary(inA, inB)
}
\arguments{
\item{inA}{path to modelA file}

\item{inB}{path to modelB file}
}
\value{
a list with results of comparison between models A and B
}
\description{
overallEvaluateSummary
}
