% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-overall-statistics.R
\name{overallStatisticsSummary}
\alias{overallStatisticsSummary}
\title{overallStatisticsSummary}
\usage{
overallStatisticsSummary(
  fileA = NULL,
  fileB = NULL,
  tNUM = getOption("RCroptest.numericThreshold"),
  tDTA = getOption("RCroptest.dateThreshold")
)
}
\arguments{
\item{fileA}{path to file (Summary.OUT) to be parsed from modelA.}

\item{fileB}{path to file (Summary.OUT) to be parsed from modelB.}

\item{tNUM}{threshold value for numeric comparison between models (default to 2).}

\item{tDTA}{threshold value for date comparsion between models (default to 0).}
}
\value{
a list with dataset from model A, modelB and warnings
}
\description{
Check if input files (Summary.OUT) from both models could be compared.
}
