% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-parser-files.R
\name{parseEvaluateFile}
\alias{parseEvaluateFile}
\title{parseEvaluateFile}
\usage{
parseEvaluateFile(inputfile)
}
\arguments{
\item{inputfile}{path to file (Evaluate.OUT) to be parsed.}
}
\value{
a list with dataset and model name
}
\description{
parseEvaluateFile
}
