% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-parser-files.R
\name{parseSummaryFile}
\alias{parseSummaryFile}
\title{parseSummaryFile}
\usage{
parseSummaryFile(inputfile)
}
\arguments{
\item{inputfile}{path to file (Summary.OUT) to be parsed.}
}
\value{
a list with model information ID and a data.fraem in JSON format.
}
\description{
parseSummaryFile
}
