% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{preProcessEvaluate}
\alias{preProcessEvaluate}
\title{preProcessEvaluate}
\usage{
preProcessEvaluate(dt)
}
\arguments{
\item{dt}{data.frame with raw Evaluate data to preprocess}
}
\value{
data.frame Summary statistics for use with evaluate comparison.
}
\description{
preProcessEvaluate
}
