% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCroptest.R
\docType{package}
\name{RCroptest}
\alias{RCroptest}
\title{\code{RCroptest} package}
\description{
RCroptest: API to process and analyse Croptest comparison
}
\details{
Provides a flexible API for files comparison. It is part of Croptest
Desktop application.
}
