% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summaryModel}
\alias{summaryModel}
\title{Show Summary model}
\usage{
summaryModel(filename)
}
\arguments{
\item{filename}{path to the Summary output file from DSSAT-CSM to be parsed.}
}
\value{
A list with version and data.frame with the parsed data from summary output file.
}
\description{
Show Summary model
}
