% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summaryStatistics}
\alias{summaryStatistics}
\title{Compute the overall summary statistics from models comparison.}
\usage{
summaryStatistics(
  fileA = NULL,
  fileB = NULL,
  tNUM = getOption("RCroptest.numericThreshold", default = 2),
  tDTA = getOption("RCroptest.dateThreshold", default = 0)
)
}
\arguments{
\item{fileA}{path to the summary file from Model A.}

\item{fileB}{path to the summary file from Model B.}

\item{tNUM}{threshold value for use with numeric calculations.}

\item{tDTA}{threshodl value for use with date calculations.}
}
\value{
a data.frame with computed summary statistics between models.
}
\description{
Compute the overall summary statistics from models comparison.
}
