% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci-models.R
\name{testCI}
\alias{testCI}
\title{Check if models are different (Sum[Tot. Above Threshold] > 0).}
\usage{
testCI(fileA, fileB, thNumeric, thDate)
}
\arguments{
\item{fileA}{path to the summary file from Model A.}

\item{fileB}{path to the summary file from Model B.}

\item{thNumeric}{threshold value for use with numeric calculations.}

\item{thDate}{threshold value for use with date calculations.}
}
\value{
Boolean value with results of comparison.
}
\description{
Check if models are different (Sum[Tot. Above Threshold] > 0).
}
\examples{
#'  testCI('C:/DSSAT48/CTWork/workA/Summary.OUT',
#'  'C:/DSSAT48/CTWork/workB/Summary.OUT',
#'  thNumeric=2,
#'  thDate=0)

}
