% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{Tot}
\alias{Tot}
\title{Tot}
\usage{
Tot(x, ...)
}
\arguments{
\item{x}{a numeric vector for which the boxplot will be constructed.}

\item{...}{a list with other parameters to pass to sum function.}
}
\value{
a number of records.
}
\description{
Compute the total number of records without NA and NaN.
}
