% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{TotOutlier}
\alias{TotOutlier}
\title{TotOutlier}
\usage{
TotOutlier(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector for which the boxplot will be constructed.}

\item{na.rm}{parameter to enable/disable removing NA and NaN.}
}
\value{
a number of outliers detected
}
\description{
Compute the total number of outliers using boxplot.stats.
}
